dlfcn

+++ 9_dlfcn/lib/add.c
@@ -1,4 +1,4 @@
-int add(int var_a, int var_b)
+float add(float var_a, float var_b)
 {
     return (var_a + var_b);
 }
