+++ 9_dlfcn/lib/div.c
@@ -1,4 +1,4 @@
-int div(int var_a, int var_b)
+float div(float var_a, float var_b)
 {
     return (var_a / var_b);
 }
