+++ 9_dlfcn/lib/mul.c
@@ -1,4 +1,4 @@
-int mul(int var_a, int var_b)
+float mul(float var_a, float var_b)
 {
     return (var_a * var_b);
 }
