+++ 9_dlfcn/lib/sub.c
@@ -1,4 +1,4 @@
-int sub(int var_a, int var_b)
+float sub(float var_a, float var_b)
 {
     return (var_a - var_b);
 }
